Ignore license agreement from plist results
Some .dmg files have a explicit license agreement that
the user needs to answer Y to be able to mount it.
hdiutil returns this on stdout and confuses xpath looking
for the mount point. The awk filter will remove anything before
it encounters a <?xml pattern which should filter out the text
from the plist information.