Redundance in defitions is no long a warn
We expect some redundancy in definitions, this allow
for multiple definitions to install a software if that
makes sense in context, exemple: jetbrains-toolbox is a common
tool necessary for java, python, data-analytics, node development.

One could just set a definition to describe jetbrains or one
could required jetbrains-toolbox in all of these languages.