Fix break of dry-run
Overriding __conditional_run leads to all operations be NOOP
which is not what we want. What we want is to the verification
run and if we would execute the do_run command we will NOOP it.

Also, we needed to fix the display messages to only NOOP
when __yellow_box 'RUN' was called. It's a hack that
overrides __yellow_box to a __magenta_box.

Another bug was when a command was skipped because it was already ran we
didn't reset the exit code which led to executing the next do_run
regardless of test.