Using a more reliable way to guess computer name
The catch with hostname is that macOS my set it using the DNS
server information and override the "Sharing" settings. When
you share a ethernet connection my router attributed one name
and when I changed laptops the same name persisted and was applied
the hostname name of the new machine. This resulted in grind thinking
that my work machine was my personal machine.